LIMS_v0_3_0/requirements.txt
Django==2.1.5
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-excel==0.0.10
django-filter==2.1.0
django-filters==0.2.1
django-tables2==2.0.3
django-tables2-column-shifter==0.5.2
lml==0.0.9
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pytz==2018.9
texttable==1.6.0
xlrd==1.2.0
xlwt==1.3.0

